requirements.txt
###
 # Specifies Python packages (and package versions) to install in order to run project.
 #
 # Execute via "pip install -r requirements.txt" while in the base project directory.
 ##


# Python SDL2 library dependencies.
pysdl2
pysdl2-dll

# System font-parsing dependencies.
fclist
